refactor: :truck: constantsをutils配下に
